fixed bug

+++ app/scripts/directives/fp-category.js
@@ -28,12 +28,15 @@ angular.module('floggitPostitsApp')
         $scope.deleteCategory = function () {
           var answer = confirm('Are you sure about deleting this category with its postits?');
           if (answer === true) {
-            for (var i = 0; i < category.postits.length; i = i + 1) {
-              dataStorage.deletePostit(currentWhiteboard.getName(), category.postits[i].id);
+            console.log('ping');
+            if (category.postits) {
+              for (var i = 0; i < category.postits.length; i = i + 1) {
+                dataStorage.deletePostit(currentWhiteboard.getName(), category.postits[i].id);
+                console.log('pang');
+              }
             }
-            dataStorage.deleteCategory(currentWhiteboard.getName(), $scope.category.id).then(function () {}).then(function () {
-              $route.reload();
-            });
+            console.log($scope.category.id);
+            dataStorage.deleteCategory(currentWhiteboard.getName(), $scope.category.id);
           }
         };
       }
